really check types in some external functions, not just DECLARE them.  fixes #15
Doing this is necessary for implementations that treat DECLARE as
promises to the compiler, rather than assertions to be checked at
runtime.

There are a number of other instances of this in external interfaces;
fixes for those can come at a later point.  The digest functions are
probably the most used bits of ironclad anyway.