fix bug with padding of 64-bit SHA variants
The SHA512/384 spec says that the final bit length of the message is to
be stored as a 128-bit (!) integer at the end of the message.  We don't
really expect anybody to be digesting 2^61+1 bytes of data, but we do
need to make sure we pad the message correctly for the length.
Fortunately, NIST's byte-level test vectors for SHA512/384 include tests
for just the cases that we were missing; add those tests and go ahead
and include their short byte-level tests for other SHA variants for
completeness's sake.

Fixes #51.